+++ IPC/Settings.cs
@@ -62,8 +62,8 @@ public static partial class WrathIPCWrapper
             RawMethod.IsCurrentJobConfiguredOn.InvokeFunc()) ??
         new Dictionary<ComboTargetTypeKeys, ComboSimplicityLevelKeys?>
         {
-            { ComboTargetTypeKeys.SingleTarget, null },
-            { ComboTargetTypeKeys.MultiTarget, null },
+            {ComboTargetTypeKeys.SingleTarget, null},
+            {ComboTargetTypeKeys.MultiTarget, null},
         };
 
     public static Dictionary<ComboTargetTypeKeys, ComboSimplicityLevelKeys?>
@@ -72,8 +72,8 @@ public static partial class WrathIPCWrapper
             RawMethod.IsCurrentJobAutoModeOn.InvokeFunc()) ??
         new Dictionary<ComboTargetTypeKeys, ComboSimplicityLevelKeys?>
         {
-            { ComboTargetTypeKeys.SingleTarget, null },
-            { ComboTargetTypeKeys.MultiTarget, null },
+            {ComboTargetTypeKeys.SingleTarget, null},
+            {ComboTargetTypeKeys.MultiTarget, null},
         };
 
     #endregion
@@ -110,6 +110,19 @@ public static partial class WrathIPCWrapper
         SafeInvokeRawMethod(() =>
             RawMethod.SetComboOptionState
                 .InvokeFunc(lease, optionName, comboState));
+
+    public static string? GetVariantParentComboName(uint jobID) =>
+        SafeInvokeRawMethod(() =>
+            RawMethod.GetVariantParentComboName.InvokeFunc(jobID));
+
+    public static List<string>? GetVariantOptionNames(uint jobID) =>
+        SafeInvokeRawMethod(() =>
+            RawMethod.GetVariantOptionNames.InvokeFunc(jobID));
+
+    public static SetResult SetVariantReadyForJob
+        (Guid lease, uint jobID, bool enabled = true) =>
+        SafeInvokeRawMethod(() =>
+            RawMethod.SetVariantReadyForJob.InvokeFunc(lease, jobID, enabled));
 
     #endregion
 }